code-rabbit fixes
