+++ apps/web/src/app/admin/students/[id]/edit/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation";
 import { StudentFormNew } from "@/features/students/components/StudentFormNew";
 
 async function getStudent(id: string) {
-	const baseUrl = `https://${process.env.VERCEL_URL}` || "http://localhost:3001";
+	const baseUrl = process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : "http://localhost:3001";
 	const response = await fetch(`${baseUrl}/api/students/${id}`, {
 		cache: "no-store",
 	});
