+++ apps/web/src/app/admin/students/assessments/[id]/edit/page.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent } from "@/components/ui/card";
 import { AssessmentForm } from "@/features/assessments/components/AssessmentForm";
 
 async function getAssessment(id: string) {
-	const baseUrl = `https://${process.env.VERCEL_URL}` || "http://localhost:3001";
+	const baseUrl = process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : "http://localhost:3001";
 	const response = await fetch(`${baseUrl}/api/assessments/${id}`, {
 		cache: "no-store",
 	});
