+++ apps/web/src/app/admin/students/enrollments/[id]/edit/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation";
 import { EnrollmentFormNew } from "@/features/enrollments/components/EnrollmentFormNew";
 
 async function getEnrollment(id: string) {
-	const baseUrl = `https://${process.env.VERCEL_URL}` || "http://localhost:3001";
+	const baseUrl = process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : "http://localhost:3001";
 	const response = await fetch(`${baseUrl}/api/enrollments/${id}`, {
 		cache: "no-store",
 	});
