+++ apps/web/src/app/admin/teachers/[id]/page.tsx
@@ -3,7 +3,7 @@ import { notFound } from "next/navigation";
 import TeacherDetailsClient from "./page-client";
 
 async function getTeacher(id: string) {
-	const baseUrl = `https://${process.env.VERCEL_URL}` || "http://localhost:3001";
+	const baseUrl = process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : "http://localhost:3001";
 	const response = await fetch(`${baseUrl}/api/teachers/${id}`, {
 		cache: "no-store",
 	});
